Add constant and start new component
