+++ verify/UISynaps.swift
@@ -63,29 +63,35 @@ extension UIColor {
 
 }
 @objcMembers public class UISynaps:  UIControl {
-    open var baseUrl = "https://verify.synaps.io/"
+	open var baseUrl = VerifyType.individual.url
     var synaps: WKWebView!
     public var status: String!
     var scriptHandler: WKScriptMessageHandler!
+
     @IBInspectable
     public var primaryColor: UIColor? {
             didSet { redraw() }
         }
+
     @IBInspectable
     public var secondaryColor: UIColor? {
             didSet { redraw() }
         }
+
     @IBInspectable
     public var tier: String? {
             didSet { redraw() }
         }
+
     @objc public var sessionId: String? {
             didSet { redraw() }
     }
+
     @IBInspectable
     public var lang: String? {
             didSet { redraw() }
     }
+
     public init(frame: CGRect, scriptHandler: WKScriptMessageHandler, type: String="individual", tier: String?=nil, lang: String?=nil, primaryColor: UIColor?=nil, secondaryColor: UIColor?=nil){
         super.init(frame: frame)
         self.setBaseUrl(type: type)
@@ -96,14 +102,17 @@ extension UIColor {
         self.scriptHandler = scriptHandler
         self.initWebView()
     }
+
     public init(frame: CGRect, scriptHandler: WKScriptMessageHandler, type: String="individual") {
         super.init(frame: frame)
         self.setBaseUrl(type: type)
         self.scriptHandler = scriptHandler
     }
-    public override init(frame: CGRect) {
+
+	public override init(frame: CGRect) {
         super.init(frame: frame)
     }
+
     public init?(coder: NSCoder, scriptHandler: WKScriptMessageHandler, type: String="individual"){
         super.init(coder: coder)
         self.setBaseUrl(type: type)
@@ -114,13 +123,15 @@ extension UIColor {
     public required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
+
     func setBaseUrl(type: String) {
         if type == "individual" {
-            self.baseUrl = "https://verify.synaps.io/"
+			self.baseUrl = VerifyType.individual.url
         } else if type == "corporate" {
-            self.baseUrl = "https://verify-v3.synaps.io/"
+            self.baseUrl = VerifyType.corporate.url
         }
     }
+
     public func initWebView() {
         let contentController = WKUserContentController();
         contentController.addUserScript(WKUserScript(source: """
@@ -143,7 +154,7 @@ extension UIColor {
         trailingAnchor.constraint(equalTo: self.synaps.trailingAnchor).isActive = true
     }
     public func redraw(){
-        var url = URL(string:self.baseUrl)
+		var url = baseUrl
         let params = [
             "session_id": self.sessionId,
             "tier": self.tier,
@@ -154,17 +165,18 @@ extension UIColor {
         ]
         for (key, val) in params {
             if val != nil {
-               url?.appendQueryItem(name: key, value: val)
+               url.appendQueryItem(name: key, value: val)
             }
         }
         if (url != nil) && (self.sessionId != nil) {
-            self.synaps.load(URLRequest(url: url!))
+            self.synaps.load(URLRequest(url: url))
         }
     }
     public func setSessionId(sessionId: String){
         self.sessionId = sessionId
     }
 }
+
 @objcMembers public class UISynapsIndividual: UISynaps {
     public  override init(frame: CGRect){
         let sh =  CustomScript()
